@@ -39,85 +39,86 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.6",
-    "@angular/cdk": "^7.3.3",
-    "@angular/common": "^7.2.6",
-    "@angular/compiler": "^7.2.6",
-    "@angular/core": "^7.2.6",
-    "@angular/forms": "^7.2.6",
-    "@angular/material": "^7.3.3",
-    "@angular/platform-browser": "^7.2.6",
-    "@angular/platform-browser-dynamic": "^7.2.6",
-    "@angular/router": "^7.2.6",
-    "@igo2/auth": "^1.4.2",
-    "@igo2/common": "^1.4.2",
-    "@igo2/context": "^1.4.2",
-    "@igo2/core": "^1.4.2",
-    "@igo2/geo": "^1.4.2",
-    "@igo2/integration": "^1.4.2",
-    "@igo2/utils": "^1.4.2",
-    "@mat-datetimepicker/core": "^3.0.0-beta.0",
+    "@angular/animations": "^10.0.3",
+    "@angular/cdk": "^10.0.1",
+    "@angular/common": "^10.0.3",
+    "@angular/compiler": "^10.0.3",
+    "@angular/core": "^10.0.3",
+    "@angular/forms": "^10.0.3",
+    "@angular/material": "^10.0.1",
+    "@angular/platform-browser": "^10.0.3",
+    "@angular/platform-browser-dynamic": "^10.0.3",
+    "@angular/router": "^10.0.3",
+    "@azure/msal-angular": "^1.1.1",
+    "@igo2/auth": "~1.5.1",
+    "@igo2/common": "~1.5.1",
+    "@igo2/context": "~1.5.1",
+    "@igo2/core": "~1.5.1",
+    "@igo2/geo": "~1.5.1",
+    "@igo2/integration": "~1.5.1",
+    "@igo2/utils": "~1.5.1",
+    "@mat-datetimepicker/core": "^4.1.0",
     "@turf/point-on-feature": "^5.1.5",
+    "angular-shepherd": "^0.7.0",
     "bowser": "^2.9.0",
     "classlist.js": "^1.1.20150312",
-    "core-js": "^2.6.5",
-    "file-saver": "^1.3.8",
+    "core-js": "^3.6.5",
+    "file-saver": "^2.0.2",
     "hammerjs": "^2.0.8",
-    "html2canvas": "1.0.0-alpha.12",
+    "html2canvas": "^1.0.0-rc.5",
     "jspdf": "^1.5.3",
     "jszip": "^3.1.5",
     "jwt-decode": "^2.2.0",
     "moment": "^2.24.0",
+    "msal": "^1.4.0",
     "natives": "^1.1.6",
     "ngx-cacheable": "^1.0.9",
-    "ol": "^5.3.0",
-    "ol-mapbox-style": "^5.0.2",
-    "proj4": "^2.5.0",
-    "rxjs": "^6.4.0",
+    "ol": "^6.3.1",
+    "ol-mapbox-style": "^6.1.2",
+    "proj4": "^2.6.2",
+    "rxjs": "^6.6.0",
     "scroll-into-view-if-needed": "^2.2.20",
     "stream": "^0.0.2",
-    "tslib": "^1.9.3",
-    "typy": "^2.0.1",
+    "tslib": "^2.0.0",
+    "typy": "^3.3.0",
     "unorm": "^1.4.1",
-    "web-animations-js": "^2.3.1",
-    "zone.js": "^0.8.29"
+    "web-animations-js": "^2.3.2",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.2",
-    "@angular-devkit/build-ng-packagr": "^0.13.2",
-    "@angular/cli": "^7.3.3",
-    "@angular/compiler-cli": "^7.2.6",
+    "@angular-devkit/build-angular": "~0.1000.1",
+    "@angular-devkit/build-ng-packagr": "~0.1000.1",
+    "@angular/cli": "^10.0.1",
+    "@angular/compiler-cli": "^10.0.3",
     "@babel/core": "^7.7.4",
     "@babel/preset-env": "^7.7.4",
     "@types/geojson": "^7946.0.6",
     "@types/hammerjs": "^2.0.36",
-    "@types/jasmine": "~2.8.6",
+    "@types/jasmine": "^3.5.11",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "angular-cli-ghpages": "^0.5.3",
-    "codelyzer": "~4.5.0",
+    "@types/node": "^14.0.22",
+    "angular-cli-ghpages": "^0.6.2",
+    "codelyzer": "^6.0.0",
     "compodoc": "0.0.41",
     "conventional-changelog-cli": "^2.0.12",
-    "gulp": "^3.9.1",
+    "gulp": "^4.0.2",
     "gulp-babel": "^8.0.0",
     "gulp-json-editor": "^2.4.2",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.0.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "^5.1.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~3.3.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
     "live-server": "^1.2.1",
-    "ng-packagr": "^3.0.0-rc.2",
+    "ng-packagr": "^10.0.0",
     "npx": "^10.2.0",
-    "protractor": "~5.4.2",
+    "protractor": "~7.0.0",
     "publish-release": "^1.6.0",
-    "scss-bundle": "^2.5.0",
-    "ts-node": "~8.0.2",
-    "tsickle": "^0.34.2",
-    "tslint": "^5.12.1",
-    "typescript": "~3.2.4",
+    "ts-node": "^8.10.2",
+    "tslint": "~6.1.0",
+    "typescript": "~3.9.6",
     "webdriver-manager": "^12.1.7",
     "webpack": "^4.41.2"
   }
