@@ -26,6 +26,6 @@ exports.config = {
     });
     jasmine
       .getEnv()
-      .addReporter(new SpecReporter({ spec: { displayStacktrace: true } }));
+      .addReporter(new SpecReporter({ spec: { displayStacktrace: 'pretty' } }));
   }
 };
