@@ -1,11 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import {
-  MatTooltipModule,
-  MatIconModule,
-  MatButtonModule,
-  MatSidenavModule
-} from '@angular/material';
+
+import { MatButtonModule } from '@angular/material/button';
+import { MatIconModule } from '@angular/material/icon';
+import { MatSidenavModule } from '@angular/material/sidenav';
+import { MatTooltipModule } from '@angular/material/tooltip';
+import { MatDialogModule } from '@angular/material/dialog';
 
 import { IgoCoreModule } from '@igo2/core';
 import {
@@ -17,7 +17,9 @@ import {
   IgoFlexibleModule,
   IgoContextMenuModule,
   IgoToolModule,
-  IgoEntityTableModule
+  IgoEntityTableModule,
+  IgoEntityTablePaginatorModule,
+  IgoInteractiveTourModule
 } from '@igo2/common';
 
 import {
@@ -41,6 +43,10 @@ import { AppToastPanelModule } from './toast-panel/toast-panel.module';
 import { AppSidenavModule } from './sidenav/sidenav.module';
 
 import { PortalComponent } from './portal.component';
+import { AppToastPanelForExpansionModule } from './toast-panel-for-expansion/toast-panel-for-expansion.module';
+
+import { WelcomeWindowComponent } from './welcome-window/welcome-window.component';
+import { IgoWelcomeWindowModule } from './welcome-window/welcome-window.module';
 
 @NgModule({
   imports: [
@@ -49,6 +55,7 @@ import { PortalComponent } from './portal.component';
     MatButtonModule,
     MatIconModule,
     MatSidenavModule,
+    MatDialogModule,
     IgoCoreModule,
     IgoFeatureModule,
     IgoImportExportModule,
@@ -67,13 +74,18 @@ import { PortalComponent } from './portal.component';
     IgoIntegrationModule,
     AppExpansionPanelModule,
     AppToastPanelModule,
+    AppToastPanelForExpansionModule,
     AppSidenavModule,
     MapOverlayModule,
     IgoContextManagerModule,
     IgoContextMapButtonModule,
-    IgoEntityTableModule
+    IgoEntityTableModule,
+    IgoEntityTablePaginatorModule,
+    IgoInteractiveTourModule,
+    IgoWelcomeWindowModule
   ],
   exports: [PortalComponent],
-  declarations: [PortalComponent]
+  declarations: [PortalComponent],
+  entryComponents: [WelcomeWindowComponent]
 })
 export class PortalModule {}
