@@ -1,14 +1,17 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import {
-  MatIconModule,
-  MatButtonModule,
-  MatSidenavModule,
-  MatTooltipModule
-} from '@angular/material';
+import { MatButtonModule } from '@angular/material/button';
+import { MatIconModule } from '@angular/material/icon';
+import { MatSidenavModule } from '@angular/material/sidenav';
+import { MatTooltipModule } from '@angular/material/tooltip';
 
 import { IgoLanguageModule } from '@igo2/core';
-import { IgoPanelModule, IgoFlexibleModule, IgoToolModule } from '@igo2/common';
+import {
+  IgoPanelModule,
+  IgoFlexibleModule,
+  IgoToolModule,
+  IgoInteractiveTourModule
+} from '@igo2/common';
 import { IgoFeatureModule } from '@igo2/geo';
 import { IgoContextManagerModule } from '@igo2/context';
 
@@ -26,7 +29,8 @@ import { SidenavComponent } from './sidenav.component';
     IgoFlexibleModule,
     IgoContextManagerModule,
     IgoToolModule,
-    IgoFeatureModule
+    IgoFeatureModule,
+    IgoInteractiveTourModule
   ],
   exports: [SidenavComponent],
   declarations: [SidenavComponent]
