@@ -1,11 +1,9 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import {
-  MatIconModule,
-  MatButtonModule,
-  MatTooltipModule,
-  MatMenuModule
-} from '@angular/material';
+import { MatButtonModule } from '@angular/material/button';
+import { MatIconModule } from '@angular/material/icon';
+import { MatMenuModule } from '@angular/material/menu';
+import { MatTooltipModule } from '@angular/material/tooltip';
 
 import { IgoPanelModule, IgoStopPropagationModule, IgoActionModule } from '@igo2/common';
 import { IgoLanguageModule } from '@igo2/core';
